move index class
