+++ src/snowflake/snowpark/modin/pandas/utils.py
@@ -839,7 +839,7 @@ def try_convert_index_to_native(index_like: Any) -> Any:
     ----------
         A pandas Index if index_like is a Snowpark pandas Index, otherwise return index_like
     """
-    from snowflake.snowpark.modin.plugin._internal.index import Index
+    from snowflake.snowpark.modin.plugin.extensions.index import Index
 
     if isinstance(index_like, Index):
         index_like = index_like.to_pandas()
