+++ src/snowflake/snowpark/modin/plugin/extensions/pd_overrides.py
@@ -40,7 +40,7 @@ from snowflake.snowpark.modin.utils import _inherit_docstrings
 if TYPE_CHECKING:  # pragma: no cover
     import csv
 
-from snowflake.snowpark.modin.plugin._internal.index import Index  # noqa: F401
+from snowflake.snowpark.modin.plugin.extensions.index import Index  # noqa: F401
 
 register_pd_accessor("Index")(Index)
